11/4 日志筛选 全局异常 日常领用完善

+++ frame/src/main/java/com/gdproj/vo/MonthSignVo.java
@@ -22,7 +22,7 @@ public class MonthSignVo {
     //实际出勤
     private Integer shouldAttendanceDays;
     //出勤率
-    private Integer attendanceRate;
+    private Double attendanceRate;
     //迟到天数
     private Integer lateDays;
     //迟到记录
